Thread communication revised, PID tuned a little, Controller output plot added

+++ src/pid.h
@@ -7,7 +7,7 @@ class PID
 {
 public:
   PID(const double &, const double &, const double &, const double & Ts=0);
-  double compute(const double &);
+  int compute(const double &);
   ~PID() = default;
 
 private:
